requirements.txt
albumentations==0.5.2
av==8.0.2
certifi==2020.12.5
click==7.1.2
cycler==0.10.0
decorator==4.4.2
easydict==1.9
efficientnet-pytorch==0.7.0
faiss==1.7.0
Flask==1.1.2
future==0.18.2
imageio==2.9.0
imgaug==0.4.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1597876339545/work
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
munch==2.5.0
natsort==7.0.1
networkx==2.5
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1596233737064/work
olefile==0.46
opencv-python==4.4.0.46
opencv-python-headless==4.4.0.46
pandas @ file:///tmp/build/80754af9/pandas_1602088135163/work
Pillow @ file:///tmp/build/80754af9/pillow_1594307298074/work
pkbar==0.5
pretrainedmodels==0.7.4
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
segmentation-models-pytorch==0.1.0
Shapely==1.7.1
sip==4.19.24
six==1.15.0
sklearn==0.0
tensorboardX==2.1
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.9.3
timm==0.3.2
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.48.2
Werkzeug==1.0.1
